factor out LoopWithShadowVarsInterface
---
